// types/db/house.ts
export interface House {
    id: number;
    nombre: string;
    Descripcion: string;
    price: number;
    Ubicacion: string;
    Calificacion: number;
    Tamaniodeconstruccion: number;
    Tamaniodeterreno: number;
    Duracion: number; // Duracion del contrato
    Tipo: string; // Renta o venta
}